Fix parsing of octal escapes
The previous parsing of octal escapes broke when the last character
in the string was an octal escape, and silently ate the character
after the octal escape otherwise. This is because it didn't reset
the position correctly.

Additionally, octal escapes were not limited to three characters,
which results in incorrect handling if the character after the
octal escape is also a valid octal character.

This fixes both issues.